Clips Controller -  Categories Controller
m-n relationship create clip done
categories controller and view added

+++ TvPlays/TvPlays/Controllers/ClipsController.cs
@@ -43,13 +43,16 @@ namespace TvPlays.Controllers
             {
                 return HttpNotFound();
             }
+
+            ViewBag.ListCategories = db.Categories.OrderBy(b => b.ID).ToList();
+
+
             return View(clips);
         }
 
         //GET: video de cada Clip
         public ActionResult VideoClip(string pathClip)
         {
-            var path = Server.MapPath("~/App_Data/Videos/");
             var file = File(pathClip, "video/mp4");
             return file;
         }
@@ -58,6 +61,7 @@ namespace TvPlays.Controllers
         public ActionResult Create()
         {
             ViewBag.UserFK = new SelectList(db.Utilizadores, "ID", "Name");
+            ViewBag.ListCategories = db.Categories.OrderBy(b => b.ID).ToList();
             return View();
         }
 
@@ -80,28 +84,57 @@ namespace TvPlays.Controllers
         //}
 
         [HttpPost]
-        public ActionResult Create(HttpPostedFileBase fileupload, ClipsDTO clips)
+        public ActionResult Create(HttpPostedFileBase fileupload, ClipsDTO clips, string[] categoriasEscolhidas)
         {
-            var user = db.Users.FirstOrDefault(u => u.UserName.Equals(User.Identity.Name));
-            var user2 = db.Utilizadores.FirstOrDefault(u => u.Email.Equals(user.Email));
 
-            if (fileupload != null)
+            Utilizadores user = db.Utilizadores.SingleOrDefault(u => u.Email.Equals(User.Identity.Name));
+
+            if (fileupload != null && user != null)
             {
                 string fileName = Path.GetFileName(fileupload.FileName);
                 string cont = fileupload.ContentType;
                 string path = Server.MapPath("~/App_Data/Videos/" + fileName);
                 fileupload.SaveAs(path);
 
-                var clip = new Clips { 
-                    TitleClip = clips.TitleClip,
-                    PathClip = path,
-                    DateClip = DateTime.Now,
-                    UserFK = user2.ID
-                };
+                /// avalia se o array com a lista das escolhas de objetos de B associados ao objeto do tipo A 
+                /// é nula, ou não.
+                /// Só poderá avanção se NÃO for nula
+                if (categoriasEscolhidas == null)
+                {
+                    ModelState.AddModelError("", "Necessita escolher pelo menos um valor de B para associar ao seu objeto de A.");
+                    // gerar a lista de objetos de B que podem ser associados a A
+                    ViewBag.ListCategories = db.Categories.OrderBy(b => b.ID).ToList();
+                    // devolver controlo à View
+                    return View(clips);
+                }
 
-                db.Clips.Add(clip);
-                db.SaveChanges();
-                return View();
+                // criar uma lista com os objetos escolhidos de B
+                List<Categories> listCategories = new List<Categories>();
+                foreach (Categories item in listCategories)
+                {
+                    //procurar o objeto de B
+                    Categories category = db.Categories.Find(item.ID);
+                    // adicioná-lo à lista
+                    listCategories.Add(category);
+                }
+
+                if (ModelState.IsValid)
+                {
+                    var clip = new Clips
+                    {
+                        TitleClip = clips.TitleClip,
+                        PathClip = path,
+                        DateClip = DateTime.Now,
+                        UserFK = user.ID,
+                        ListCategories = listCategories
+                    };
+
+                    db.Clips.Add(clip);
+                    db.SaveChanges();
+                    return View();
+                }
+
+                return View(clips);
             }
             return View();
         }
@@ -127,13 +160,23 @@ namespace TvPlays.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID,TimeClip,TitleClip,DateClip,PathClip,UserFK")] Clips clips)
+        public ActionResult Edit([Bind(Include = "ID,TitleClip,DateClip,PathClip,UserFK")] Clips clips)
         {
-            if (ModelState.IsValid)
+            Utilizadores user = db.Utilizadores.SingleOrDefault(u => u.Email.Equals(User.Identity.Name));
+            Clips clip = db.Clips.Find(clips.ID);
+            
+
+            //Verificar se o utilizador Autenticado e se o mesmo é o dono do Clip, se for autoriza-se a alteração 
+            if (user.ID.Equals(clip.UserFK))
             {
-                db.Entry(clips).State = EntityState.Modified;
+                clip.TitleClip = clips.TitleClip;
+                db.Entry(clip).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                //return RedirectToAction("Index");
+            }
+            else
+            {
+                //Erro porque o comment nao e dele
             }
             ViewBag.UserFK = new SelectList(db.Utilizadores, "ID", "Name", clips.UserFK);
             return View(clips);
@@ -159,10 +202,36 @@ namespace TvPlays.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            //Ir buscar o User e o Clip a base de dados
+            Utilizadores user = db.Utilizadores.SingleOrDefault(u => u.Email.Equals(User.Identity.Name));
             Clips clips = db.Clips.Find(id);
-            db.Clips.Remove(clips);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+
+            //Se for Admin pode apagar qualquer Clip antes da verificacao do User == null && comment == null 
+            //porque o admin nao e um Utilizador normal da aplicação
+            if (User.IsInRole("Admin"))
+            {
+                db.Clips.Remove(clips);
+                db.SaveChanges();
+            }
+
+            //Verifica se tanto o user como o clip existem e retorna 404 se nao encontrar algum dos dois
+            if (user == null || clips == null)
+            {
+                return HttpNotFound();
+            }
+
+            //Verificar se o utilizador Autenticado e mesmo o dono do Clips, se for autoriza-se e concretiza-se a remoção
+            if (user.ID.Equals(clips.UserFK))
+            {
+                db.Clips.Remove(clips);
+                db.SaveChanges();
+                //return RedirectToAction("Index");
+            }
+            else
+            {
+                //Erro porque o comment nao e dele
+            }
+            return View(clips);
         }
 
         protected override void Dispose(bool disposing)
